pyproject.toml
[project]
name = "data-flywheel"
version = "0.1.0"
description = "Self-improving ML platform with closed-loop data flywheel — on-premises, CV-first, domain-agnostic"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    # Data Version Control
    "dvc[s3]>=3.55",
    # Data Validation
    "cleanlab>=2.7",
    "cleanvision>=0.3.6",
    # ML Framework
    "torch>=2.6",
    "torchvision>=0.21",
    # Experiment Tracking
    "mlflow>=3.10",
    # Orchestration
    "prefect>=3.0",
    # Configuration
    "pydantic-settings>=2.7",
    # Storage
    "boto3>=1.35",
    # Serving
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "gunicorn>=22.0",
    "python-multipart>=0.0.9",
    # Utilities
    "Pillow>=11.0",
    "numpy>=2.0",
    "pandas>=2.2",
    # Monitoring
    "prometheus-client>=0.21",
    "prometheus-fastapi-instrumentator>=7.0",
    "evidently>=0.5",
    # System metrics (required by mlflow.enable_system_metrics_logging)
    "psutil>=6.0",
    "nvidia-ml-py>=12.0",
    # Redis (multi-worker reload sync)
    "redis>=5.0",
    # Active Learning (Label Studio bridge)
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "httpx>=0.27",
    "pytest-playwright>=0.6.2",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "ANN", # flake8-annotations (enforce type hints)
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (requires Docker services)",
    "e2e: End-to-end pipeline tests",
    "playwright: Browser-based E2E tests (requires Playwright + Docker)",
]
